Correctly handle conditional jumps
From QEMUs side we need to be more careful in handling conditional
branches for PC-relative code (system mode).  We now have three forms
of branches: pcrelative, condtional (might be taken), fallthrough (if
conditional is not taken).

This is needed to manage the saved "base" pc of the TB in QEMU.

Signed-off-by: Anton Johansson <anjo@rev.ng>